requirements.txt
atomicwrites==1.2.0
attrs==18.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
comtypes==1.1.7
idna==2.7
more-itertools==4.3.0
numpy==1.15.1
opencv-python==3.4.2.17
Pillow==5.2.0
pluggy==0.7.1
py==1.5.4
pycparser==2.19
pyhooked==0.8.1
pypiwin32==223
pytesseract==0.2.4
pytest==3.7.2
pywin32==223
rabird.core==0.4.1
six==1.11.0
urllib3==1.23
win32core==221.36
win32gui==221.6
winappdbg==1.5
winiobinary==0.0.6
